[draft] test xAB
