+++ tests/plugins/org.polarsys.capella.test.diagram.tools.ju/src/org/polarsys/capella/test/diagram/tools/ju/tb/xABTitleBlockTestCase.java
@@ -1,0 +1,22 @@
+package org.polarsys.capella.test.diagram.tools.ju.tb;
+
+import org.polarsys.capella.core.model.helpers.BlockArchitectureExt;
+import org.polarsys.capella.test.diagram.common.ju.context.CommonDiagram;
+import org.polarsys.capella.test.diagram.common.ju.context.XABDiagram;
+
+public abstract class xABTitleBlockTestCase extends AbstractTitleBlockTestCase {
+  protected String diagramName;
+  protected BlockArchitectureExt.Type type;
+
+  public xABTitleBlockTestCase() {
+    initData();
+  }
+
+  @Override
+  protected CommonDiagram initDiagram() {
+    return XABDiagram.openDiagram(context, diagramName, type);
+  }
+
+  public abstract void initData();
+
+}
